Try to show messages according to current channelActionCreators
Create a constant and an action creator for clicking a channel and let the reducer react to that.
Make a record for Channel and create a interface with the id of the current channel and all channels.
Create a ChannelBarItemContainer so the callback delete and click props can be passed directly to the ChannelBarItem component.